requirements.txt
streamlit==1.26.0
Pillow>=7.1.0,<10
numpy==1.26.0
mediapipe>=0.10.20  
opencv-python-headless==4.8.1.78